backend/requirements.txt
fastapi==0.111.0
numpy==2.0.0
opencv_python==4.8.0.76
opencv_python_headless==4.10.0.82
Pillow==10.0.0
Pillow==10.3.0
rembg==2.0.57
scikit_learn==1.3.2
scipy==1.13.1
skimage==0.0
uvicorn==0.30.1
